pyproject.toml
[project]
name = "modbus-rtu-client"
version = "0.1.0"
description = "ModBusRTUClient is a simple ModBus RTU client library"
authors = [
    {name = "CQ Zha",email = "cqzha_2201@hotmail.co.uk"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bitarray (>=3.6.1,<4.0.0)",
    "pyserial (>=3.5,<4.0)"
]

[tool.poetry]
packages = [{include = "modbus_rtu_client", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
